+++ kleides_mfa/signals.py
@@ -1,0 +1,4 @@
+from django.dispatch import Signal
+
+mfa_added = Signal()
+mfa_removed = Signal()